Remove ra_ResponseWindow from RAN configs
The ra_ResponseWindow is the (time) window during which the RAN has to
send Msg2 in response to a Msg1 (preamble). It can be selected, but the
spec says that it should not be more than 10ms.

Changes in oai/openairinterface5g!3521 lead to the gNB more often
violate this time window. In fact, previously there was a specific
configuration in the RAN that avoided this violation; for harmonization
and simplification purposes, this is being removed.  On the other hand,
for SCS=1, an ra_ResponseWindow of 4 (10 slots=5ms) as currently
configured is overly restrictive, as the L1 slot ahead time is already
5 or 6 slots, so we already lose more than half the time for answering
with Msg2.

oai/openairinterface5g!3521 reworks this to have the RAN automatically
select the ra_ResponseWindow. The current default for all configs is 5,
so removing these values from the config will allow
oai/openairinterface5g!3521 to pass pipeline OAI-CN5G-COTS-UE-Test,
before the configuration parameter will be selected automatically.

See also https://jenkins-oai.eurecom.fr/job/OAI-CN5G-COTS-UE-Test/7850/